In computer class, the first assignment was to write a program to print the first 100 Fibonacci numbers. Instead, I wrote a program that would steal passwords of students. My teacher gave me an A

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,19 +7,19 @@
 */
 int main(void)
 {
-	long long int prev, curr, next;
+	long int prev, curr, next;
 	int counter;
 
 	prev = 1;
 	curr = 2;
 	next = prev + curr;
 
-	printf("%lld, %lld", prev, curr);
+	printf("%ld, %ld", prev, curr);
 
 	for (counter = 2; counter < 50; counter++)
 	{
 		next = prev + curr;
-		printf(", %lld", next);
+		printf(", %ld", next);
 		prev = curr;
 		curr = next;
 	}
